+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
    * Added support for `multipart/form-data` content type in PUT and PATCH requests
    * Added support for object relationships
+   * Added variables `$environment` (string), `$request` (PSR-7 ServerRequestInterface|null) and `$uri` (PSR-7 Uri|null) to be used in `setup.php`
 
 # v1.7.34
 ## 06/14/2022